refactor: TS & Bun, reorganized project
